requirements.txt
click==7.1.2
confluent-kafka==1.6.1
dependency-injector==4.31.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
MarkupSafe==1.1.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
